Separate out chunks of code and create and put them to new files to organize code
